Refactor variables/files from  PeerPrep to PeerCode
