+++ example-docker-compose.env
@@ -5,13 +5,13 @@ SERVER_PORT=5000
 MONGO_ROOT_USER=root
 MONGO_ROOT_PASSWORD=password
 # Database details
-MONGO_PEERPREP_HOST_NAME=localhost
-MONGO_PEERPREP_DATABASE_NAME=peerprep_db
-MONGO_PEERPREP_DATABASE_USER=peerprep_db_user
-MONGO_PEERPREP_DATABASE_PASSWORD=peerprep_db_password
+MONGO_PEERCODE_HOST_NAME=localhost
+MONGO_PEERCODE_DATABASE_NAME=peercode_db
+MONGO_PEERCODE_DATABASE_USER=peercode_db_user
+MONGO_PEERCODE_DATABASE_PASSWORD=peercode_db_password
 
 # Kafka details
-KAFKA_SERVER_NAME=localhost:19092  # localhost:19092 for dev, otherwise peerprep_kafka:9092
+KAFKA_SERVER_NAME=localhost:19092  # localhost:19092 for dev, otherwise peercode_kafka:9092
 KAFKA_GROUP_SERVER=kafka_group_server 
 KAFKA_GROUP_QUESTION_SERVICE=kafka_group_question_service
 KAFKA_GROUP_LEETCODE_SERVICE=kafka_group_leetcode_service
