+++ mongo-init.js
@@ -1,12 +1,12 @@
-db = new Mongo().getDB(process.env.MONGO_PEERPREP_DATABASE_NAME);
+db = new Mongo().getDB(process.env.MONGO_PEERCODE_DATABASE_NAME);
 
 db.createUser({
-  user: process.env.MONGO_PEERPREP_DATABASE_USER,
-  pwd: process.env.MONGO_PEERPREP_DATABASE_PASSWORD,
+  user: process.env.MONGO_PEERCODE_DATABASE_USER,
+  pwd: process.env.MONGO_PEERCODE_DATABASE_PASSWORD,
   roles: [
     {
       role: "readWrite", // cannot be 'root'
-      db: process.env.MONGO_PEERPREP_DATABASE_NAME,
+      db: process.env.MONGO_PEERCODE_DATABASE_NAME,
     },
   ],
 });
